Fix delete route if parkiraliste doesn't exists

+++ backend/controllers/parkiraliste/DeleteParkiralisteController.ts
@@ -1,6 +1,7 @@
 import { BaseController } from '../BaseController';
 import { IRequest, IResponse } from '../../interfaces/network';
 import { ParkiralisteRepo } from '../../repos/ParkiralisteRepo';
+import { TvrtkaRepo } from '../../repos/TvrtkaRepo';
 
 export class DeleteParkiralisteController extends BaseController {
   executeImpl = async (
@@ -15,6 +16,18 @@ export class DeleteParkiralisteController extends BaseController {
 
     if (idParkiraliste < 1) {
       return this.clientError(res, ['Id parkirališta mora biti pozitivan']);
+    }
+
+    const idTvrtka = await ParkiralisteRepo.getIdTvrtkaFromIdParkiraliste(
+      idParkiraliste
+    );
+
+    if (
+      !idTvrtka ||
+      req.session.user.idRacun !=
+        (await TvrtkaRepo.getIdRacunFromIdTvrtka(idTvrtka))
+    ) {
+      return this.forbidden(res, null);
     }
 
     await ParkiralisteRepo.deleteByIdParkiraliste(idParkiraliste);
